refactor(MenuService): dependency injection for menu logic in VendingMachine and Menu domains
